fix: ddev

+++ .ddev/web-entrypoint.d/farah.sh
@@ -2,6 +2,6 @@
 
 set -eu
 
-if [ ! -e /var/www/vendor ]; then
+if [ ! -e /var/www/vendor ] && [ ! -L /var/www/vendor ]; then
     ln -s /var/www/html/vendor /var/www/vendor
 fi
